Migrate some tests
